requirements.txt
aiohttp==3.9.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
arclet-alconna==1.8.43
arclet-alconna-tools==0.7.12
argcomplete==3.6.3
attrs==25.4.0
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
exceptiongroup==1.3.1
fastapi==0.124.4
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.0
jinja2==3.1.6
loguru==0.7.3
markupsafe==3.0.3
multidict==6.7.0
nepattern==0.7.7
nonebot-adapter-qq==1.6.6
nonebot-plugin-alconna==0.60.3
nonebot-plugin-authrespond==1.8
nonebot-plugin-localstore==0.7.4
nonebot-plugin-session==0.3.2
nonebot-plugin-status==0.9.0
nonebot-plugin-waiter==0.8.1
nonebot2==2.4.4
nonestorage==0.1.0
outcome==1.3.0.post0
packaging==25.0
pillow==10.3.0
pip==24.0
pipx==1.8.0
platformdirs==4.5.1
propcache==0.4.1
psutil==6.1.1
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.7.4
pydantic-core==2.18.4
pygtrie==2.5.0
pymysql==1.1.2
pysocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.3
requests==2.32.5
selenium==4.22.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.45
sqlmodel==0.0.27
starlette==0.50.0
steam==1.4.4
strenum==0.4.15
tarina==0.7.3
trio==0.32.0
trio-websocket==0.12.2
typing-extensions==4.15.0
urllib3==2.6.2
userpath==1.9.2
uvicorn==0.38.0
uvloop==0.22.1
vdf==3.4
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
